covid etl/requirements.txt
pandas~=2.2.3
apache-airflow~=2.10.5
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
openpyxl~=3.1.5
numpy~=2.2.3
SQLAlchemy~=1.4.54